File restructring
