+++ .flake8
@@ -1,0 +1,2 @@
+[flake8]
+ignore = E741,W504,C812,W605,E266
