+++ src/app_main.py
@@ -6,8 +6,8 @@ from aiokafka import AIOKafkaConsumer
 from fastapi import FastAPI, HTTPException
 
 # ML modules
-from ml_modules.bert_classifier import BertClassifier
-from ml_modules.rule_based_clustering import RuleBasedClassifier
+from src.ml_modules.bert_classifier import BertClassifier
+from src.ml_modules.rule_based_clustering import RuleBasedClassifier
 from pydantic import BaseModel
 
 # import ml_modules.run_zsc as zsc
